SB-1, i modified the package name of the itemWriter class and modified the writer.xml file
