+++ src/main/resources/writer.xml
@@ -25,7 +25,7 @@
 		<!-- We need a custom setter to set values to Query param by getting values Bank object -->
 		<property name="itemPreparedStatementSetter">
 			<bean
-				class="com.ab.batch.writer.BankItemPSSetter" />
+				class="com.ab.writer.BankItemPSSetter" />
 		</property>
 	</bean>
